backend/requirements.txt
# FastAPI & Web
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
pydantic>=2.10.0
pydantic-settings>=2.12.0

# LangChain
langchain>=0.3.14
langchain-core>=0.3.29
langchain-openai>=0.3.0
langchain-google-genai>=2.0.8
langchain-anthropic>=0.3.1
langchain-community>=0.3.14
langchain-text-splitters>=0.3.5
langsmith>=0.2.11

# LLM SDKs
openai>=1.59.0
google-generativeai>=0.8.4
anthropic>=0.42.0

# Auth
PyJWT>=2.10.0
passlib>=1.7.4
bcrypt>=5.0.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
redis>=5.0.0
qdrant-client>=1.12.0

# ML/AI
torch>=2.0.0
transformers>=4.40.0
sentence-transformers>=2.2.0
FlagEmbedding>=1.3.0
accelerate>=1.0.0
safetensors>=0.4.0
peft>=0.10.0
tiktoken>=0.5.0

# File processing
aiofiles>=23.0.0
python-docx>=1.0.0
PyMuPDF>=1.24.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
filetype>=1.2.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
orjson>=3.9.0
tenacity>=8.2.0
